10.15.3

+++ sudo/sudo/lib/util/sha2.c
@@ -1,5 +1,7 @@
 /*
- * Copyright (c) 2013-2015 Todd C. Miller <Todd.Miller@courtesan.com>
+ * SPDX-License-Identifier: ISC
+ *
+ * Copyright (c) 2013-2015 Todd C. Miller <Todd.Miller@sudo.ws>
  *
  * Permission to use, copy, modify, and distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -12,6 +14,11 @@
  * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ */
+
+/*
+ * This is an open source non-commercial project. Dear PVS-Studio, please check it.
+ * PVS-Studio Static Code Analyzer for C, C++ and C#: http://www.viva64.com
  */
 
 /*
@@ -255,7 +262,7 @@ SHA256Update(SHA2_CTX *ctx, const uint8_t *data, size_t len)
 	size_t i = 0, j;
 
 	j = (size_t)((ctx->count[0] >> 3) & (SHA256_BLOCK_LENGTH - 1));
-	ctx->count[0] += (len << 3);
+	ctx->count[0] += ((uint64_t)len << 3);
 	if ((j + len) > SHA256_BLOCK_LENGTH - 1) {
 		memcpy(&ctx->buffer[j], data, (i = SHA256_BLOCK_LENGTH - j));
 		SHA256Transform(ctx->state.st32, ctx->buffer);
@@ -466,8 +473,8 @@ SHA512Update(SHA2_CTX *ctx, const uint8_t *data, size_t len)
 	size_t i = 0, j;
 
 	j = (size_t)((ctx->count[0] >> 3) & (SHA512_BLOCK_LENGTH - 1));
-	ctx->count[0] += (len << 3);
-	if (ctx->count[0] < (len << 3))
+	ctx->count[0] += ((uint64_t)len << 3);
+	if (ctx->count[0] < ((uint64_t)len << 3))
 		ctx->count[1]++;
 	if ((j + len) > SHA512_BLOCK_LENGTH - 1) {
 		memcpy(&ctx->buffer[j], data, (i = SHA512_BLOCK_LENGTH - j));
